backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pydriller==2.5.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.8
anthropic==0.7.8
python-dotenv==1.0.0